Add .profile to source control
This .profile file is unchanged from what is provided in Ubuntu 18.04
by default. When setting up a new machine, it may be necessary to remove
the existing file (a major downside of putting it under source control).

We could add a .bash_profile, but that wouldn't be used as long as this
file exists.