Upgrade from a.vim to alternate-lite
The alternate-lite plugin is a simplification/improvement of the
original a.vim script.

Your current system with a.vim was broken when you had all of an h,
hpp, and cpp file. If you started from an cpp file you would get stuck
only switching between the h and cpp, and if you started from an hpp you
would get stuck between the hpp and cpp, and if you started from an h
file you would get stuck between the h and hpp.

The problem is that once a.vim has loaded one alternate it checks for
an existing buffer and always prefers that to loading a file from disk.
See the bestFile logic in the AlternateFile function.